refactor(web): extract UpdateBanner into its own component (#66)
